+++ lib/src/features/remote_localization/domain/contracts/remote_localization_connector.dart
@@ -2,6 +2,7 @@ import 'dart:ui' show Locale;
 
 import '../entities/remote_localization_payload.dart';
 import '../entities/remote_localization_cache_snapshot.dart';
+import '../entities/remote_localization_version.dart';
 import '../entities/remote_update_descriptor.dart';
 
 abstract interface class RemoteLocalizationConnector {