+++ lib/src/features/remote_localization/domain/repositories/remote_localization_repository.dart
@@ -3,7 +3,6 @@ import 'dart:ui' show Locale;
 import '../entities/remote_localization_cache_snapshot.dart';
 import '../entities/remote_localization_payload.dart';
 import '../entities/remote_localization_result.dart';
-import '../entities/remote_update_descriptor.dart';
 import '../entities/remote_localization_version.dart';
 
 abstract interface class RemoteLocalizationRepository {